refactor server connection and endpoints to own file
